chore(*): refactor build and testing process
