+++ .babelrc
@@ -1,10 +1,5 @@
 {
     "presets": [
-        "es2015-loose", "stage-1"
-    ],
-    "plugins": [
-        "syntax-flow",
-        "transform-flow-strip-types",
-        "transform-decorators-legacy"
+        "es2015", "stage-1"
     ]
 }
