+++ gulpfile.babel.js
@@ -1,53 +1,52 @@
-import gulp from "gulp";
-import paths from "vinyl-paths";
-import del from "del";
-import run from "run-sequence";
-import jasmine from "gulp-jasmine";
-import reporters from "jasmine-reporters";
-import istanbul from "gulp-istanbul";
-import tsb from "gulp-tsb";
-import tsconfig from "./tsconfig.json";
-import remapIstanbul from "remap-istanbul/lib/gulpRemapIstanbul";
+import gulp from 'gulp';
+import paths from 'vinyl-paths';
+import del from 'del';
+import run from 'run-sequence';
+import jasmine from 'gulp-jasmine';
+import reporters from 'jasmine-reporters';
+import istanbul from 'gulp-istanbul';
+import typescript from 'typescript';
+import gts from 'gulp-typescript';
+import sourcemaps from 'gulp-sourcemaps';
+import merge from "merge2";
+import tsconfig from './tsconfig.json';
+import remapIstanbul from 'remap-istanbul/lib/gulpRemapIstanbul';
 
-const typescriptSources = tsconfig.compilerOptions.rootDir + "/**/*.ts";
+const typescriptSources = tsconfig.compilerOptions.rootDir + '/**/*.ts';
 const typescriptOutput = tsconfig.compilerOptions.outDir;
-const typescriptOutputTesting = "test/dist";
-const testSuites = [ "test/unit/**/*.ts" ];
-const testSuitesDist = [ "test/dist/test/unit/**/*.js" ];
+const testSuites = [ 'dist/test/unit/**/*.js' ];
 
-const tsc = tsb.create(tsconfig.compilerOptions);
-const tscTesting = tsb.create(Object.assign({}, tsconfig.compilerOptions, {
-    outDir: typescriptOutputTesting,
-    target: "es5",
-    moduleResolution: "node",
-    declaration: false
-}));
+const tsc = gts(Object.assign({ typescript: typescript }, tsconfig.compilerOptions));
 
-gulp.task("clean", () => gulp.src([ typescriptOutput ]).pipe(paths(del)));
-gulp.task("clean-test", () => gulp.src(typescriptOutputTesting).pipe(paths(del)));
-gulp.task("build", [ "clean" ], () => gulp.src(typescriptSources).pipe(tsc()).pipe(gulp.dest(typescriptOutput)));
-gulp.task("build-test", [ "clean-test" ], () => gulp.src(typescriptSources).pipe(tscTesting()).pipe(gulp.dest(`${typescriptOutputTesting}/src`)));
-gulp.task("instrument-test", [ "build-test" ], () =>
-    gulp.src(`${typescriptOutputTesting}/src/**/*.js`)
+gulp.task('clean', done => gulp.src([ typescriptOutput ]).pipe(paths(del)));
+gulp.task('build', [ 'clean' ], done => {
+    let stream = gulp.src(typescriptSources).pipe(sourcemaps.init()).pipe(tsc);
+    return merge([
+        stream.js.pipe(sourcemaps.write('.', { mapSources: sourcePath => '../' + sourcePath }))
+            .pipe(gulp.dest(typescriptOutput)),
+        stream.dts.pipe(gulp.dest(typescriptOutput))
+    ]);
+});
+gulp.task('instrument-test', [ 'build' ], done => {
+    return gulp.src(typescriptOutput + '/main/**/*.js')
         .pipe(istanbul())
-        .pipe(istanbul.hookRequire())
-);
-gulp.task("build-unit-test", ['instrument-test'], () => gulp.src(testSuites).pipe(tscTesting()).pipe(gulp.dest(typescriptOutputTesting)));
-gulp.task('test', ['build-unit-test'], () =>
-    gulp.src(testSuitesDist)
+        .pipe(istanbul.hookRequire());
+});
+gulp.task('test', [ 'instrument-test' ], done => {
+    return gulp.src(testSuites)
         .pipe(jasmine())
         .pipe(istanbul.writeReports({
-            reporters: ['json']
-        }))
-);
-gulp.task('remap-istanbul', () =>
-    gulp.src('coverage/coverage-final.json')
+            reporters: [ 'json' ]
+        }));
+});
+gulp.task('remap-istanbul', done => {
+    return gulp.src('coverage/coverage-final.json')
         .pipe(remapIstanbul({
             reports: {
                 'lcovonly': 'coverage/lcov.info',
                 'json': 'coverage/coverage.json',
                 'html': 'coverage/html-report'
             }
-        }))
-);
-gulp.task("default", () => run("build", "test"));
+        }));
+});
+gulp.task('default', done => run('test'));
