+++ src/main/invocation-handler.ts
