--- typings.json
@@ -1,9 +1,0 @@
-{
-  "name": "aurelia-property-injection",
-  "version": "0.4.0",
-  "main": "dist/index.d.ts",
-  "dependencies": {},
-  "globalDevDependencies": {
-    "jasmine": "registry:dt/jasmine#2.2.0+20160621224255"
-  }
-}
